+++ wazo_confd/plugins/line/plugin.py
@@ -3,6 +3,7 @@
 
 from wazo_provd_client import Client as ProvdClient
 
+from xivo_dao.resources.endpoint_custom import dao as endpoint_custom_dao
 from xivo_dao.resources.endpoint_sccp import dao as endpoint_sccp_dao
 from xivo_dao.resources.line import dao as line_dao
 from xivo_dao.resources.endpoint_sip import dao as sip_dao
@@ -11,9 +12,11 @@ from xivo_dao.resources.pjsip_transport import dao as transport_dao
 from wazo_confd.plugins.endpoint_sip.service import (
     build_endpoint_service as build_endpoint_sip_service,
 )
+from wazo_confd.plugins.line_endpoint.service import build_service_custom
 from wazo_confd.plugins.line_endpoint.service import build_service_sip
 from wazo_confd.plugins.line_endpoint.service import build_service_sccp
 from wazo_confd.plugins.endpoint_sccp.service import build_service as build_endpoint_sccp_service
+from wazo_confd.plugins.endpoint_custom.service import build_service as build_endpoint_custom_service
 
 from .resource import LineItem, LineList
 from .service import build_service
@@ -30,8 +33,10 @@ class Plugin:
         token_changed_subscribe(provd_client.set_token)
 
         service = build_service(provd_client)
+        endpoint_custom_service = build_endpoint_custom_service()
         endpoint_sccp_service = build_endpoint_sccp_service()
         endpoint_sip_service = build_endpoint_sip_service(provd_client, pjsip_doc)
+        line_endpoint_custom_association_service = build_service_custom(provd_client)
         line_endpoint_sip_association_service = build_service_sip(provd_client)
         line_endpoint_sccp_association_service = build_service_sccp(provd_client)
 
@@ -46,10 +51,13 @@ class Plugin:
             '/lines',
             resource_class_args=(
                 service,
+                endpoint_custom_service,
                 endpoint_sip_service,
+                line_endpoint_custom_association_service,
                 line_endpoint_sip_association_service,
                 line_endpoint_sccp_association_service,
                 endpoint_sccp_service,
+                endpoint_custom_dao,
                 endpoint_sccp_dao,
                 line_dao,
                 sip_dao,